Tugas 6

+++ main/views.py
@@ -1,5 +1,5 @@
 import datetime
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseNotFound
 from django.urls import reverse
 from django.core import serializers
 from django.shortcuts import render
@@ -11,6 +11,7 @@ from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import csrf_exempt
 
 @login_required(login_url='/login')
 def show_main(request):
@@ -22,6 +23,7 @@ def show_main(request):
         'class': 'PBP E',
         'name': request.user.username,
         'items': items,
+        'items_count': len(Item.objects.filter(user=request.user)),
         'last_login': request.COOKIES['last_login'],
     }
 
@@ -100,7 +102,30 @@ def edit_product(request, id):
     context = {'form': form}
     return render(request, "edit_product.html", context)
 
+@csrf_exempt
 def delete_product(request, id):
-    product = Item.objects.get(pk = id)
-    product.delete()
-    return HttpResponseRedirect(reverse('main:show_main'))
+    if request.method == 'DELETE':
+        product = Item.objects.get(pk=id)
+        product.delete()
+        
+        return HttpResponse(b"REMOVED", status=201)
+    return HttpResponseNotFound()
+
+def get_product_json(request):
+    product_item = Item.objects.filter(user=request.user)
+    return HttpResponse(serializers.serialize('json', product_item))
+
+@csrf_exempt
+def add_product_ajax(request):
+    if request.method == 'POST':
+        name = request.POST.get("name")
+        price = request.POST.get("price")
+        amount = request.POST.get("amount")
+        description = request.POST.get("description")
+        user = request.user
+
+        new_product = Item(name=name, price=price, amount=amount, description=description, user=user)
+        new_product.save()
+        return HttpResponse(b"CREATED", status=201)
+
+    return HttpResponseNotFound()
